Worked over BaHaMAS tests adding new ones
In particular some commands of the real code (e.g. sbatch, less) were
inhibited to avoid jobs submissions as well as interruptions.

+++ Tests/AuxiliaryFunctions.bash
@@ -4,7 +4,7 @@
 
 function CreateTestsFolderStructure()
 {
-    cd $BaHaMAS_testsFolder || exit -2
+    cd "$BaHaMAS_testsFolder" || exit -2
     mkdir -p "${testFolder}${testParametersPath}"
     mkdir -p "${testFolder}/Rational_Approximations"
     cp "${BaHaMAS_testsFolderAuxFiles}/fakeApprox" "${testFolder}/Rational_Approximations"
@@ -15,25 +15,50 @@ function CreateTestsFolderStructure()
 
 function MakeTestPreliminaryOperations()
 {
+    #Always go at betafolder level and then in case cd elsewhere
+    cd "${testFolder}${testParametersPath}" || exit -2
+    #Always delete everything inside (paranoic check?!)
+    [ $(pwd) = "${testFolder}${testParametersPath}" ]  && rm -rf "$betaFolder" "betas"
     #Always use completed file and then in case overwrite
     cp "${BaHaMAS_testsFolderAuxFiles}/fakeBetas" "${testFolder}${testParametersPath}/betas"
-    #Always go at betafolder level and then in case cd elsewhere
-    cd "${testFolder}${testParametersPath}"
 
     case "$1" in
         completeBetasFile* )
             cp "${BaHaMAS_testsFolderAuxFiles}/fakeBetasToBeCompleted" "${testFolder}${testParametersPath}/betas"
             ;;
+        liststatus* )
+            local file
+            mkdir "$betaFolder"
+            for file in fakeStdOutput fakeInput fakeOutput; do
+                cp "${BaHaMAS_testsFolderAuxFiles}/${file}" "${testFolder}${testParametersPath}/${betaFolder}"
+            done
+                ;;
         * )
             ;;
     esac
 }
 
+function InhibitBaHaMASCommands()
+{
+    function less(){ cecho -d "less $@"; }
+    function sbatch(){ cecho -d "sbatch $@"; }
+    function squeue(){ builtin squeue "$@"; } #Change this?!
+    export -f less sbatch
+}
+
 function RunBaHaMASInTestMode()
 {
-    printf "\n\n$(date)\n  Running:\n    ${BaHaMAS_command} $@\n\n" >> $logFile
-    #In subshell to exclude any potential variables conflict and in test mode!
-    ( BaHaMAS_testModeOn='TRUE' ${BaHaMAS_command} "$@" >> $logFile )
+    printf "\n===============================\n" >> $logFile
+    printf " $(date)\n" >> $logFile
+    printf "===============================\n" >> $logFile
+    printf "Running:\n    ${BaHaMAS_command} $@\n\n" >> $logFile
+    # NOTE: Here we run BaHaMAS in subshell to exclude any potential variables conflict.
+    #       Moreover we activate the test mode defining a variable before running it and
+    #       we inhibit some commands in order to avoid job summission. Observe also that
+    #       at the moment there are no options or options value which have a space in them
+    #       and, hence, we can pass $@ to BaHaMAS instead of "$@" in order to word split
+    #       the string passed to this function. This will break also spaces inside options!
+    ( InhibitBaHaMASCommands; BaHaMAS_testModeOn='TRUE' ${BaHaMAS_command} $@ >> $logFile )
     if [ $? -eq 0 ]; then
         return 0
     else
@@ -49,7 +74,7 @@ function RunTest()
     if [ $reportLevel -eq 3 ]; then
         printf -v stringTest "%-38s" "__${testName}$(cecho -d bb)_"
         stringTest="${stringTest// /.}"
-        cecho -n bb "  $(printf '%+2s' ${testsRun})/$(printf '%-2s' ${#availableTests[@]})" emph "${stringTest//_/ }"
+        cecho -n bb "  $(printf '%+2s' ${testsRun})/$(printf '%-2s' ${#testsToBeRun[@]})" emph "${stringTest//_/ }"
     fi
     RunBaHaMASInTestMode "$@"
     if [ $? -eq 0 ]; then
@@ -69,12 +94,12 @@ function RunTest()
 function PrintTestsReport()
 {
     local indentation name
-    indentation='    '
+    indentation='          '
     if [ $reportLevel -ge 1 ]; then
         cecho bb "\n${indentation}==============================\n"\
-              lp "${indentation}   Run " emph "$(printf '%2d' $testsRun)" " tests: "\
+              lp "${indentation}   Run " emph "$(printf '%2d' $testsRun)" " test(s): "\
               lg "$(printf '%2d' $testsPassed) passed\n"\
-              lr "${indentation}                 $(printf '%2d' $testsFailed) failed\n"\
+              lr "${indentation}                   $(printf '%2d' $testsFailed) failed\n"\
               bb "${indentation}=============================="
     fi
     if [ $reportLevel -ge 2 ]; then
@@ -99,8 +124,8 @@ function PrintTestsReport()
 function CleanTestsEnvironment()
 {
     local name
-    cd $BaHaMAS_testsFolder || exit -2
-    [ $reportLevel -eq 3 ] && cecho bb "\n In $(pwd):"
+    cd "$BaHaMAS_testsFolder" || exit -2
+    [ $reportLevel -eq 3 ] && cecho bb " In $(pwd):"
     for name in "${listOfAuxiliaryFilesAndFolders[@]}"; do
         if [ $testsFailed -ne 0 ] && [ $name = $logFile ]; then
             continue
@@ -112,7 +137,7 @@ function CleanTestsEnvironment()
         elif ls "$name" 1>/dev/null 2>&1; then
             cecho lr "   Error in $FUNCNAME: " emph "$name" " neither file or directory, leaving it!"; continue
         fi
-        rm -rf "$name"
+        [ $(pwd) = "$BaHaMAS_testsFolder" ] && rm -rf "$name"
     done
     cecho ''
 }
